backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
web3==6.0.0
eth-account>=0.8.0 